:bath: Refactor tests to reduce duplication
